+++ src/Mingine.Physics/Simulator.fs
@@ -9,14 +9,14 @@ open FSharp.Collections
 // workaround for F# compiler internal error kekw
 let private vecOrigin = { x = 0.<_>; y = 0.<_> } // Vec2.origin
 
-let guardFloatInstability<[<Measure>] 'u> =
+let private guardFloatInstability<[<Measure>] 'u> =
     mapFloatTyped<'u> (fun value ->
         if Double.IsNaN value || Math.Abs value = infinity then
             0.
         else
             value)
 
-let guardForceTorqueInstability (f, t) =
+let inline private guardForceTorqueInstability (f, t) =
     Vec2.map guardFloatInstability f, guardFloatInstability t
 
 let calcForcesAndTorques (gObj: PhysicsObj) (timeStep: float<s>) =
@@ -33,46 +33,46 @@ let calcForcesAndTorques (gObj: PhysicsObj) (timeStep: float<s>) =
         // sum all forces and torques
         |> Array.reduce (fun (f1, t1) (f2, t2) -> (f1 + f2, t1 + t2))
 
-let updateObjectPos gObj timeStep =
+let updateObjectPos pObj timeStep =
     // update transform from last tick info
     let newPos =
-        gObj.pos
-        + (gObj.velocity * timeStep)
-        + (gObj.accel * 0.5 * (timeStep * timeStep))
+        pObj.pos
+        + (pObj.velocity * timeStep)
+        + (pObj.accel * 0.5 * (timeStep * timeStep))
 
     let newAngle =
-        gObj.angle
-        + (gObj.angVelocity * timeStep)
-        + (gObj.angAccel * 0.5 * (timeStep * timeStep))
+        pObj.angle
+        + (pObj.angVelocity * timeStep)
+        + (pObj.angAccel * 0.5 * (timeStep * timeStep))
 
     // calculate forces and torques
     let force, torque =
         calcForcesAndTorques
-            {gObj with
+            {pObj with
                 pos = newPos
                 angle = newAngle}
             timeStep
 
     // update this tick acceleration
     // radians are dimensionless so are not part of the torque nor moment of inertia units, so add them in with *1rad
-    let newAccel = force / gObj.mass
+    let newAccel = force / pObj.mass
 
     let newAngAccel =
-        torque / gObj.momentOfInertia * 1.<rad>
+        torque / pObj.momentOfInertia * 1.<rad>
 
-    let avgAccel = (gObj.accel + newAccel) / 2.
+    let avgAccel = (pObj.accel + newAccel) / 2.
 
     let avgAngAccel =
-        (gObj.angAccel + newAngAccel) / 2.
+        (pObj.angAccel + newAngAccel) / 2.
 
     // update this tick velocity
     let newVelocity =
-        gObj.velocity + (avgAccel * timeStep)
+        pObj.velocity + (avgAccel * timeStep)
 
     let newAngVelocity =
-        gObj.angVelocity + (avgAngAccel * timeStep)
+        pObj.angVelocity + (avgAngAccel * timeStep)
 
-    {gObj with
+    {pObj with
         pos = newPos
         accel = newAccel
         velocity = newVelocity
@@ -80,3 +80,5 @@ let updateObjectPos gObj timeStep =
         angle = newAngle
         angAccel = newAngAccel
         angVelocity = newAngVelocity}
+
+let impulse (force: Vec2<_>) pObj = { pObj with accel = pObj.accel + (force / pObj.mass) }